@@ -1,6 +1,5 @@
 // натсройки для переадресации на другие компоненты
-
-window.Vue = require('vue')
+import Vue from "vue"
 import VueRouter from 'vue-router';
 
 Vue.use(VueRouter);
