@@ -1,12 +1,5 @@
 module.exports = {
-  "presets": [
-    "@babel/preset-env"
-  ],
-  "env": {
-    "test": {
-      "presets": [
-        ["@babel/preset-env", { "targets": { "node": "current" } }]
-      ]
-    }
-  }
+  presets: [
+    '@vue/cli-plugin-babel/preset'
+  ]
 }
