@@ -148,7 +148,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
    @import '../../sass/autorization.sass'
 
