@@ -455,7 +455,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
    @import '../../sass/dataForProfile.sass'
 
