@@ -118,7 +118,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
    @import '../../sass/mainPage.sass'
 
