@@ -89,7 +89,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
    @import '../../sass/modalDeleteProfile.sass'
 
