@@ -629,7 +629,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
     @import '../sass/dashBoard.sass'
    
