@@ -444,7 +444,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
     @import '../../sass/dashBarChart.sass'
 
