@@ -323,7 +323,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
     @import '../../sass/dashButton.sass'
 
