@@ -597,7 +597,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
     @import '../../sass/dashCSvg.sass'
 
