@@ -136,7 +136,7 @@ export default {
 
 </script>
 
- <style lang="sass" > 
+ <style lang="scss" > 
   
       @import '../../sass/dashFilter.sass'
 
