@@ -840,7 +840,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
     @import '../../sass/dashGraph.sass'
 
