@@ -714,8 +714,8 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
 
-    @import '../../sass/dashGuntt.sass';
+    @import '../../sass/dashGuntt.sass'
 
 </style>
