@@ -1005,7 +1005,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
      @import '../../sass/dashLineChart.sass'
     
