@@ -1882,7 +1882,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
      @import '../../sass/dashMultiLine.sass'
     
