@@ -416,7 +416,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
     @import '../../sass/dashPieChart.sass'
 
