@@ -388,7 +388,7 @@ export default {
  
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
     @import '../../sass/dashSelect.sass'
 
