@@ -232,6 +232,6 @@ export default {
 };
 </script>
 
-<style lang="sass" >
+<style lang="scss" >
 @import '../../sass/dashTile.sass'
 </style>
