@@ -1557,7 +1557,7 @@ export default {
 }
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
 
         @import '../sass/dashPanelBoard.sass'
 </style>
