@@ -183,7 +183,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
     @import '../sass/dashSettings.sass'
 
