@@ -24,7 +24,7 @@ export default {
 }
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
    @import '../sass/footer.sass'
 
