@@ -189,8 +189,8 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
 
-    @import '../../sass/formEdit.sass';
+    @import '../../sass/formEdit.sass'
 
 </style>
