@@ -190,7 +190,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
 
     // настройки стилей в файле formEdit.sass
 
