@@ -91,8 +91,8 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
 
-    @import '../../sass/modalChooseElement.sass';
+    @import '../../sass/modalChooseElement.sass'
 
 </style>
