@@ -427,7 +427,7 @@ export default {
 }
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
 
 
     @import '../sass/modalCreate.sass'
