@@ -324,7 +324,7 @@ export default {
 }
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
 
      @import '../sass/modalCreateSearch.sass'
     
