@@ -271,7 +271,7 @@ export default {
 
 </script>
 
-<style lang="sass" >
+<style lang="scss" >
   
   @import '../sass/modalExin.sass'
    
