@@ -277,7 +277,7 @@ export default {
 
 </script>
 
-<style lang="sass"  > 
+<style lang="scss"  > 
   
   @import '../sass/modalPaper.sass'
    
