@@ -212,7 +212,7 @@ export default {
 }
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
 
      @import '../sass/modalReports.sass'
     
