@@ -384,7 +384,7 @@ export default {
 </script>
 
 
-<style lang="sass"> 
+<style lang="scss"> 
   
     @import '../sass/modalScheduler.sass'
    
