@@ -118,7 +118,7 @@ export default {
 </script>
 
 
-<style lang="sass" scoped> 
+<style lang="scss" scoped> 
   
     @import '../sass/modalSearch.sass'
    
