@@ -1440,6 +1440,6 @@ export default {
 }
 </script>
 
-<style lang="sass">
+<style lang="scss">
   @import '../sass/modalSettings.sass'  
 </style>
