@@ -816,7 +816,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
    @import '../sass/papers.sass'
    
