@@ -563,7 +563,7 @@ export default {
 
 </script>
 
-<style lang="sass" > 
+<style lang="scss" > 
   
    @import '../sass/reports.sass'
    
